RecommendationAgent/requirements.txt
langchain 
langchain_openai
langchain_community
openai 
requests 
pandas
python-dotenv
streamlit